\name{bayesian-model-zoo-package}
\alias{bayesian-model-zoo-package}
\alias{bayesian-model-zoo}
\docType{package}
\title{
\packageTitle{bayesian-model-zoo}
}
\description{
\packageDescription{bayesian-model-zoo}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{bayesian-model-zoo}
\packageIndices{bayesian-model-zoo}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{bayesian-model-zoo}

Maintainer: \packageMaintainer{bayesian-model-zoo}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
