% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesianmodelzoo-package.R
\docType{package}
\name{bayesianmodelzoo-package}
\alias{bayesianmodelzoo-package}
\alias{bayesmodelzoo}
\title{Bayesian Model Zoo}
\description{
\if{html}{
   \figure{stanlogo.png}{options: width="50px" alt="mc-stan.org"}
   \emph{Stan Development Team}
}

The \pkg{bayesmodelzoo} package is an appendage to the \pkg{rstan} package
}
