% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changepoint.R
\name{changepoint}
\alias{changepoint}
\title{Perform a changepoint analysis}
\usage{
changepoint(x, chains = 2, iter = 1000, allpars = FALSE,
  cores = max(parallel::detectCores() - 1, 1), ...)
}
\arguments{
\item{x}{the time series to perform the analysis on}

\item{chains}{the number of chains to use which defaults to 2}

\item{iter}{the number of samples to pull which defaults to 1000}

\item{allpars}{decides if all parameters should be returned or not. FALSE which is
the default only gives the mu's and the sigmas and the tau}

\item{cores}{the number of cores to use which defaults to max(parallel::detectCores()-1, 1)}

\item{...}{other arguments passed to the sampling method in rstan}
}
\value{
the stanfit from the sampled changepoint model
}
\description{
Samples the changepoint model using Stan and returns the stanfit object as a
result.
}
\examples{
\dontrun{
library(bayesplot)
aa<-changepoint(c(rnorm(50, 100, 10), rnorm(50, 150, 10)))
mcmc_combo(as.array(aa), regex_pars = "tau")
}
}
