% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changepoint.R
\name{distOverlap}
\alias{distOverlap}
\title{Quantify distribution overlap}
\usage{
distOverlap(x1, x2, breaks = NULL)
}
\arguments{
\item{x1}{samples from distribution 1}

\item{x2}{samples from distribution 2}

\item{breaks}{manually specify the breaks to use which if set to NULL (the default) will be estimated by histogram}
}
\value{
the overlap in percent represented as a number between 0 and 1
}
\description{
Measures the amount of overlap between two sampled numerical distributions x1
and x2. The return value states the degree of overlap. If 1 then 100 percent
is overlapped meanwhile a 0 indicates no overlap.
}
\examples{
distOverlap(rnorm(10, 10, 5), rnorm(10, 5, 5))
}
