% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negbinom.R
\name{negbinom}
\alias{negbinom}
\title{Bayesian negative binomial regression}
\usage{
negbinom(y, x, chains = 2, iter = 1000, allpars = FALSE,
  cores = max(parallel::detectCores() - 1, 1), ...)
}
\arguments{
\item{y}{the response variable to uses}

\item{x}{the matrix of covariates to use}

\item{chains}{the number of chains to use which defaults to 2}

\item{iter}{the number of samples to pull which defaults to 1000}

\item{allpars}{decides if all parameters should be returned or not. FALSE which is
the default only gives the mu's, beta's, phi and predicted y_rep}

\item{cores}{the number of cores to use which defaults to max(parallel::detectCores()-1, 1)}

\item{...}{other arguments passed to the sampling method in rstan}
}
\value{
the stanfit from the sampled model
}
\description{
This implements a Bayesian negative binomial regression by using the 2nd
parameterization from the Stan manual.
}
\examples{
\dontrun{
  data<-list(N=nrow(mtcars), y=mtcars$gear, X=mtcars[, c('drat', 'am')])
  sfit<-negbinom(data$y, data$X, iter = 500, chains=2, cores=2, allpars = TRUE)
  library(bayesplot)
  mcmc_combo(as.array(sfit), regex_pars = "tau")
  yrep<-as.matrix(sfit)
  ppc_dens_overlay(data$y, yrep[1:50,grep("y_", colnames(yrep))])
}
}
