% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppcaard.R
\name{ppcaard}
\alias{ppcaard}
\title{Perform a probabilistic PCA with ARD}
\usage{
ppcaard(xdf, k, chains = 2, iter = 500,
  cores = max(parallel::detectCores() - 1, 1), ...)
}
\arguments{
\item{xdf}{the data set to perform the analysis on}

\item{k}{the number of latent factors to consider}

\item{chains}{the number of chains to use which defaults to 2}

\item{iter}{the number of samples to pull which defaults to 1000}

\item{cores}{the number of cores to use which defaults to max(parallel::detectCores()-1, 1)}

\item{...}{other arguments passed to the sampling method in rstan}
}
\value{
the stanfit from the sampled changepoint model
}
\description{
This function performs a probabilistic principal component analysis with
automatic relevance determination.
}
\examples{
library(ggplot2)
library(mvtnorm)
tmpdf <- mvtnorm::rmvnorm(400, sigma = matrix(c(1,0.5,0.5,1),2,2))
qplot(tmpdf[,1], tmpdf[,2])
}
